requirements-dev.txt
# Development dependencies
-r requirements.txt

# Pre-commit hooks
pre-commit>=3.6.0

# Code formatting
black>=24.1.0
isort>=5.13.0

# Linting
flake8>=7.0.0
flake8-docstrings>=1.7.0
flake8-bugbear>=24.1.0
pylint>=3.0.0
pylint-django>=2.5.0

# Type checking
mypy>=1.8.0
django-stubs>=4.2.0
djangorestframework-stubs>=3.14.0

# Security
bandit>=1.7.0

# Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
coverage>=7.4.0
